Redis connection changes for go1.0.3 compatibility
We now create new connections for every redis request.  Since the
connections are cached, this should not impact performance with network
overhead.  This helps us by not having the requirement of having locks
on the shared connection.  Furthermore, we can process more concurrent
requests this way.  Finally, we don't run up on the issue of go1.0.3
processing global scopes slightly differently.